requirements.txt
flask==2.3.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
flask-pymongo==2.3.0
python-dotenv==1.0.0
email-validator==2.0.0
flask-mail==0.9.1
pymongo==4.5.0